multilayer update example 1
